requirements.txt
streamlit==1.42.0
plotly==6.0.0
google-cloud-bigquery==3.27.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
requests==2.32.3
pandas==2.2.3
python-dotenv==1.0.1
gspread==6.1.4
db-dtypes==1.3.1
